+++ packages/browserless/src/http.js
@@ -3,10 +3,11 @@ import { expandReferences } from '@openfn/language-common/util';
 
 /**
  * Make a Browserless-authenticated HTTP request operation.
+ * This operation always sends requests to the configured Browserless 'baseUrl'.
  * @public
  * @function
  * @param {string} method - HTTP method (e.g. 'GET', 'POST').
- * @param {string} url - Path or URL to request (relative paths are joined to the configured `baseUrl`).
+ * @param {string} path - API Path (e.g. 'pdf, 'content) joined to the configured 'baseUrl'.
  * @param {object} options - Request options (body, headers, query, parseAs).
  * @state {HttpState}
  * @returns {Operation} an OpenFn operation returning a state with `data` and `response`.